Compiler warnings removed

+++ externals/Nandrad/src/NANDRAD_SolarLoadsDistributionModel.cpp
@@ -41,7 +41,7 @@ void SolarLoadsDistributionModel::checkParameters() const {
 	// check radiation load fractions
 	for (unsigned int i=0; i<4; ++i)
 		m_para[P_RadiationLoadFractionZone+i].checkedValue(
-					NANDRAD::KeywordList::Keyword("SolarLoadsDistributionModel::para_t", P_RadiationLoadFractionZone+i),
+					NANDRAD::KeywordList::Keyword("SolarLoadsDistributionModel::para_t", P_RadiationLoadFractionZone+(int)i),
 					"---", "---", 0, true, 1, true,
 					"Radiation load fraction must be between 0 and 1.");
